modified footer, changed icon, commented payment gateway
