+++ index.html
@@ -10,7 +10,7 @@
 		/>
 		<link
 			rel="shortcut icon"
-			href="./assets/logo/logo.png"
+			href="./assets/logo/favicon.ico"
 			type="image/x-icon"
 		/>
 
@@ -21,9 +21,9 @@
 		/>
 		<meta property="og:description" content="Bring your collection to life" />
 		<meta property="og:type" content="website" />
-		<meta property="og:url" content="https://github.com/PaulleDemon" />
+		<meta property="og:url" content="https://figurebase.us" />
 		<!--Replace with the current website url-->
-		<meta property="og:image" content="./assets/logo/logo.png" />
+		<meta property="og:image" content="./assets/logo/logo-new.png" />
 
 		<!-- <link rel="stylesheet" href="../../tailwind-css/tailwind-runtime.css" /> -->
 		<link rel="stylesheet" href="css\tailwind-build.css" />
@@ -44,7 +44,7 @@
 		>
 			<a class="tw-h-[50px] tw-w-[50px] tw-p-[4px]" href="">
 				<img
-					src="./assets/logo/logo.png"
+					src="./assets/logo/logo-new.png"
 					alt="logo"
 					class="tw-object tw-h-full tw-w-full"
 				/>
@@ -651,23 +651,20 @@
 						<i class="bi bi-plus tw-ml-auto tw-font-semibold"></i>
 					</div>
 					<div class="content tw-w-full max-w-lg tw-overflow-auto">
-						Be among the first to experience the magic of Figure Base. Transform
-						your static collection into an interactive experience that grows
-						more engaging every day.
-
+						Preorder your product today with no upfront payment. You will not be charged until we confirm your order closer to the official launch date.
 						<div class="tw-p-8 tw-rounded-lg tw-shadow-lg tw-w-full max-w-lg">
-							<h2 class="tw-text-2xl tw-font-bold tw-mb-6 tw-text-center">Payment Form</h2>
+							<h2 class="tw-text-2xl tw-font-bold tw-mb-6 tw-text-center">Pre-order</h2>
 							<form id="payment-form" class="tw-flex tw-flex-col tw-items-center tw-gap-4">
 								<div class="tw-w-full">
 									<label class="tw-text-gray-700">Name</label>
-									<input type="text" id="name" name="name" required class="tw-w-full tw-p-2 tw-border tw-border-gray-300 tw-rounded-md">
+									<input type="text" id="name" name="name" required class="tw-w-full tw-p-2 tw-border tw-border-gray-300 tw-rounded-md" style="color: black;">
 								</div>
 								<div class="tw-w-full">
 									<label class=" tw-text-gray-700">Email</label>
-									<input type="email" id="email" name="email" required class="tw-w-full tw-p-2 tw-border tw-border-gray-300 tw-rounded-md">
+									<input type="email" id="email" name="email" required class="tw-w-full tw-p-2 tw-border tw-border-gray-300 tw-rounded-md" style="color: black;">
 								</div>
 								<button type="submit" class="btn tw-w-500 tw-text-white tw-p-2 tw-rounded-md">
-									Proceed to Payment
+									Add to pre-order list
 								</button>
 							</form>
 						</div>
@@ -678,7 +675,7 @@
 			<div class="tw-mt-20 tw-flex tw-flex-col tw-place-items-center tw-gap-4">
 				<div class="tw-text-3xl max-md:tw-text-2xl">Still have questions?</div>
 				<a
-					href="http://"
+					href="mailto:contact@figurebase.us"
 					class="btn !tw-rounded-full !tw-border-[1px] !tw-border-solid !tw-border-gray-300 !tw-bg-transparent tw-transition-colors tw-duration-[0.3s]"
 				>
 					Contact
@@ -742,7 +739,7 @@
 				class="tw-flex tw-h-full tw-w-[250px] tw-flex-col tw-place-items-center tw-gap-6 max-md:tw-w-full"
 			>
 				<img
-					src="./assets/logo/logo.png"
+					src="./assets/logo/logo-new.png"
 					alt="logo"
 					srcset=""
 					class="tw-max-w-[120px]"
@@ -783,14 +780,18 @@
 				</div>
 			</div> -->
 
-			<div class="tw-flex tw-h-full tw-w-[250px] tw-flex-col tw-gap-4">
+			<div class="tw-flex tw-h-full tw-w-[300px] tw-flex-col tw-gap-4">
 				<h2 class="tw-text-3xl max-md:tw-text-xl">Resources</h2>
 				<div class="tw-flex tw-flex-col tw-gap-3 max-md:tw-text-sm">
-					<a href="" class="footer-link">About us</a>
-					<a href="" class="footer-link">FAQ</a>
-					<a href="" class="footer-link">Contact Us</a>
-					<a href="" class="footer-link">Blogs</a>
-					<a href="" class="footer-link">Privacy policy</a>
+					<p class="text-sm">
+						&copy; 2025 FigureBase.
+						All rights reserved.
+					</p>
+					<!-- <a href="" class="footer-link">About us</a>
+					<a href="" class="footer-link">FAQ</a> -->
+					<!-- <a href="" class="footer-link">Contact Us</a> -->
+					<!-- <a href="" class="footer-link">Blogs</a>
+					<a href="" class="footer-link">Privacy policy</a> -->
 				</div>
 			</div>
 		</footer>
@@ -809,10 +810,12 @@
 		referrerpolicy="no-referrer"
 	></script>
 
+	<!-- locally deployed variables -->
 	<script type="module" src="./db-config.js"></script>
-	<script type="module" src="./store-record.js"></script>
-	<script src="./index.js"></script>
 
+	<!-- other scripts -->
+	<script type="module" src="./api/store-record.js"></script>
+	<script src="./index.js"></script>
 	<script src="https://js.stripe.com/v3/"></script>
 
 </html>
